reorganise webserver auth config code
