+++ cmd/openem-ingestor-app/app.go
@@ -91,7 +91,7 @@ func (a *App) startup(ctx context.Context) {
 	a.taskqueue.Startup()
 
 	go func(port int) {
-		ingestor, err := webserver.NewIngestorWebServer(a.version, &a.taskqueue, a.config.Auth)
+		ingestor, err := webserver.NewIngestorWebServer(a.version, &a.taskqueue, a.config.WebServerAuth)
 		if err != nil {
 			panic(err)
 		}
